Reorganize models into one file per class. Run mypy on the codebase and fix validation errors.
