+++ config.example.py
@@ -3,12 +3,14 @@ import tempfile
 from pathlib import Path
 
 # Add your botpassword and login here:
+from typing import List
+
 username = ""
 password = ""
 
 # Global settings
 wiki_user = "User:Username"  # Change this to your username
-list_of_allowed_aliases = [] # Add elements like this ["API"]
+list_of_allowed_aliases: List[str] = []  # Add elements like this ["API"]
 logging.basicConfig(level=logging.WARNING)
 version = "0.2"  # Don't touch this.
 wd_prefix = "http://www.wikidata.org/entity/"
