+++ requirements.txt
@@ -2,3 +2,4 @@ console-menu
 git+git://github.com/LeMyst/WikibaseIntegrator@v0.12.0.dev5#egg=wikibaseintegrator
 rich~=10.9.0
 SPARQLWrapper~=1.8.5
+pydantic
