+++ src/models/scholarly_articles.py
@@ -1,83 +1,87 @@
 import logging
 
-from wikibaseintegrator.wbi_helpers import execute_sparql_query
+from wikibaseintegrator.wbi_helpers import execute_sparql_query  # type: ignore
 
 import config
 from src.helpers.cleaning import strip_bad_chars
 from src.helpers.console import console
 from src.models.suggestion import Suggestion
 from src.models.task import Task
-from src.models.wikidata import Items, Item
-
-
-def build_query(suggestion: Suggestion = None,
-                search_string: str = None,
-                task: Task = None,
-                cirrussearch_parameters: str = None):
-    if suggestion is None:
-        raise ValueError("suggestion was None")
-    if search_string is None:
-        raise ValueError("search_string was None")
-    if task is None:
-        raise ValueError("task was None")
-    if cirrussearch_parameters is None:
-        raise ValueError("cirrussearch_parameters was None")
-    # This query uses https://www.w3.org/TR/sparql11-property-paths/ to
-    # find subjects that are subclass of one another up to 3 hops away
-    # This query also uses the https://www.mediawiki.org/wiki/Wikidata_Query_Service/User_Manual/MWAPI
-    # which has a hardcoded limit of 10,000 items so you will never get more matches than that
-    # This query use regex to match beginning, middle and end of the label of matched items
-    # The replacing lines should match the similar python replacements in cleaning.py
-    # The replacing with "\\\\\\\\" becomes "\\\\" after leaving python and then it works in
-    # SPARQL where it becomes "\\" and thus match a single backslash
-    return (f"""
-        #{config.user_agent}
-        SELECT DISTINCT ?item ?itemLabel 
-        WHERE {{
-          hint:Query hint:optimizer "None".
-          BIND(STR('{cirrussearch_parameters} \"{search_string}\"') as ?search_string)
-          SERVICE wikibase:mwapi {{
-            bd:serviceParam wikibase:api "Search";
-                            wikibase:endpoint "www.wikidata.org";
-                            mwapi:srsearch ?search_string.
-            ?title wikibase:apiOutput mwapi:title. 
-          }}
-          BIND(IRI(CONCAT(STR(wd:), ?title)) AS ?item)
-          ?item rdfs:label ?label.
-          BIND(REPLACE(LCASE(?label), ",", "") as ?label1)
-          BIND(REPLACE(?label1, ":", "") as ?label2)
-          BIND(REPLACE(?label2, ";", "") as ?label3)          
-          BIND(REPLACE(?label3, "\\\\(", "") as ?label4)
-          BIND(REPLACE(?label4, "\\\\)", "") as ?label5)
-          BIND(REPLACE(?label5, "\\\\[", "") as ?label6)
-          BIND(REPLACE(?label6, "\\\\]", "") as ?label7)
-          BIND(REPLACE(?label7, "\\\\\\\\", "") as ?label8)
-          BIND(?label8 as ?cleaned_label)
-          FILTER(CONTAINS(?cleaned_label, ' {search_string.lower()} '@{task.language_code.value}) || 
-                 REGEX(?cleaned_label, '.* {search_string.lower()}$'@{task.language_code.value}) ||
-                 REGEX(?cleaned_label, '^{search_string.lower()} .*'@{task.language_code.value}))
-          MINUS {{?item wdt:P921/wdt:P279 wd:{suggestion.item.id}. }}
-          MINUS {{?item wdt:P921/wdt:P279/wdt:P279 wd:{suggestion.item.id}. }}
-          MINUS {{?item wdt:P921/wdt:P279/wdt:P279/wdt:P279 wd:{suggestion.item.id}. }}
-          SERVICE wikibase:label {{ bd:serviceParam wikibase:language "en". }}
-        }}
-        """)
-
-
-def process_results(results):
-    items = []
-    for item_json in results["results"]["bindings"]:
-        logging.debug(f"item_json:{item_json}")
-        item = Item(json=item_json)
-        items.append(item)
-    return items
+from src.models.wikidata.item import Item
+from src.models.wikidata.items import Items
 
 
 class ScholarlyArticleItems(Items):
     """This supports both published peer reviewed articles and preprints"""
+
     def fetch_based_on_label(self,
                              suggestion: Suggestion = None,
                              task: Task = None):
+        def build_query(suggestion: Suggestion = None,
+                        search_string: str = None,
+                        task: Task = None,
+                        cirrussearch_parameters: str = None):
+            # TODO refactor
+            if suggestion is None:
+                raise ValueError("suggestion was None")
+            if search_string is None:
+                raise ValueError("search_string was None")
+            if task is None:
+                raise ValueError("task was None")
+            if task.language_code is None:
+                raise ValueError("task.language_code was None")
+            if cirrussearch_parameters is None:
+                raise ValueError("cirrussearch_parameters was None")
+            # This query uses https://www.w3.org/TR/sparql11-property-paths/ to
+            # find subjects that are subclass of one another up to 3 hops away
+            # This query also uses the https://www.mediawiki.org/wiki/Wikidata_Query_Service/User_Manual/MWAPI
+            # which has a hardcoded limit of 10,000 items so you will never get more matches than that
+            # This query use regex to match beginning, middle and end of the label of matched items
+            # The replacing lines should match the similar python replacements in cleaning.py
+            # The replacing with "\\\\\\\\" becomes "\\\\" after leaving python and then it works in
+            # SPARQL where it becomes "\\" and thus match a single backslash
+            return (f"""
+                #{config.user_agent}
+                SELECT DISTINCT ?item ?itemLabel 
+                WHERE {{
+                  hint:Query hint:optimizer "None".
+                  BIND(STR('{cirrussearch_parameters} \"{search_string}\"') as ?search_string)
+                  SERVICE wikibase:mwapi {{
+                    bd:serviceParam wikibase:api "Search";
+                                    wikibase:endpoint "www.wikidata.org";
+                                    mwapi:srsearch ?search_string.
+                    ?title wikibase:apiOutput mwapi:title. 
+                  }}
+                  BIND(IRI(CONCAT(STR(wd:), ?title)) AS ?item)
+                  ?item rdfs:label ?label.
+                  BIND(REPLACE(LCASE(?label), ",", "") as ?label1)
+                  BIND(REPLACE(?label1, ":", "") as ?label2)
+                  BIND(REPLACE(?label2, ";", "") as ?label3)          
+                  BIND(REPLACE(?label3, "\\\\(", "") as ?label4)
+                  BIND(REPLACE(?label4, "\\\\)", "") as ?label5)
+                  BIND(REPLACE(?label5, "\\\\[", "") as ?label6)
+                  BIND(REPLACE(?label6, "\\\\]", "") as ?label7)
+                  BIND(REPLACE(?label7, "\\\\\\\\", "") as ?label8)
+                  BIND(?label8 as ?cleaned_label)
+                  FILTER(CONTAINS(?cleaned_label, ' {search_string.lower()} '@{task.language_code.value}) || 
+                         REGEX(?cleaned_label, '.* {search_string.lower()}$'@{task.language_code.value}) ||
+                         REGEX(?cleaned_label, '^{search_string.lower()} .*'@{task.language_code.value}))
+                  MINUS {{?item wdt:P921/wdt:P279 wd:{suggestion.item.id}. }}
+                  MINUS {{?item wdt:P921/wdt:P279/wdt:P279 wd:{suggestion.item.id}. }}
+                  MINUS {{?item wdt:P921/wdt:P279/wdt:P279/wdt:P279 wd:{suggestion.item.id}. }}
+                  SERVICE wikibase:label {{ bd:serviceParam wikibase:language "en". }}
+                }}
+                """)
+
+        def process_results(results):
+            # TODO refactor
+            items = []
+            for item_json in results["results"]["bindings"]:
+                logging.debug(f"item_json:{item_json}")
+                item = Item(json=item_json)
+                items.append(item)
+            return items
+
         # logger = logging.getLogger(__name__)
         if suggestion is None:
             raise ValueError("suggestion was None")
