+++ src/models/thesis.py
@@ -1,11 +1,12 @@
 import logging
 
-from wikibaseintegrator.wbi_helpers import execute_sparql_query
+from wikibaseintegrator.wbi_helpers import execute_sparql_query  # type: ignore
 
 from src.helpers.console import console
 from src.models.suggestion import Suggestion
 from src.models.task import Task
-from src.models.wikidata import Items, Item
+from src.models.wikidata.item import Item
+from src.models.wikidata.items import Items
 
 # There were ~16.000 thesis' in WD when this was written
 
@@ -22,6 +23,8 @@ class ThesisItems(Items):
                             "supported yet for this task.")
         if task is None:
             raise ValueError("task was None")
+        if task.language_code is None:
+            raise ValueError("task.language_code was None")
         # Fetch all items maching the search strings
         self.list = []
         for search_string in suggestion.search_strings:
